+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.3] - 2022-08-09
+### Added
+- A isPublic column to OwnedEntity
+- A isAdmin column to User
+
+## [0.0.2] - 2022-08-07
 ### Added
 - `/auth/login` route to get a token
 - A Note route
@@ -28,4 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A Book route
 - A Reading route
 
+[Unreleased]: https://github.com/taylorhmorris/reading-log-back/compare/v1.0.0...HEAD
+[0.0.2]: https://github.com/taylorhmorris/reading-log-back/compare/v0.0.2...v0.0.3
+[0.0.2]: https://github.com/taylorhmorris/reading-log-back/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/taylorhmorris/reading-log-back/releases/tag/v0.0.1